+++ src/game/types.ts
@@ -1,6 +1,7 @@
 /**
  * Shared types and interfaces
  * @copyright 2025 LeanderKafemann. All rights reserved.
+ * @version 1.1.0
  */
 
 export interface LeaderboardEntry {
@@ -14,18 +15,4 @@ export interface GameConfig {
   canvasHeight: number;
   winScore: number;
   paddleSpeed: number;
-}
-
-export enum AbilityType {
-  SMASH = 'smash',
-  SHIELD = 'shield',
-  SPEED_BOOST = 'speedBoost'
-}
-
-export interface Ability {
-  type: AbilityType;
-  name: string;
-  cooldown: number;
-  duration: number;
-  key: string;
 }